Move volume command into own file
